v8bindings: Enable the fix of cross origin window name visibility
Enables the fix of cross origin window name visibility in
production.

This fix was once applied back in 2017, but got reverted due to
a bug in youtube.com (not because the fix was wrong).  We've been
disabling the fix since then.  The bug in youtube.com was fixed
many years ago, so nothing should prevent us to enable the fix now.

https://chromestatus.com/metrics/feature/timeline/popularity/1872
shows the affected cases of this fix as 0.045% of page loads.
So, we shouldn't hit the same severe trouble like the one of
youtube.com in the past.

Bug: 538562
Change-Id: Ia28090493c2eb1f67d65de52f64690bb755ed38a