NativeIO: Add deleteAfterClose option to open
The deleteAfterClose option adds a way to mark a file for removal after
closure. This is useful to manage the lifecycle of temporary files, e.g.
when implementing a cache that should not be persisted.

This change adds the "main" path, creating or opening an existing file
with the deleteAfterClose option. For existing files, the previously
written contents should be readable. For both cases, the file should be
deleted once it's closed.

Design doc: https://docs.google.com/document/d/1MyVykFuWtCmziN8t5C115fsbD1d3hwr5qO5gk3l09CY/edit?usp=sharing
Bug: 1155570
Change-Id: I7bb03a237e753283696984896e9df25d52bdb786