Pass initiator frame token for location.reload()
Previously, we were treating navigations initiated by
location.reload() in the renderer as navigations without initiator. It
seems better to let the initiator be the window being reloaded (as it
happens, for example, for a reload initiated by a refresh meta tag).
This has the effect that in the browser process we correctly inherit
the policy container policies.

Change-Id: Id0951b3602ea889fee0f465afbda645b6b1d3cbd
Fixed: 1185145
Bug: 1185145