CSP: Fix report-to for frame-ancestors violation
Content-Security-Policy 'frame-ancestors' checks are performed in the
browser during the navigation. Violation reports, however, are sent
through Blink. When using the reporting API (via the 'report-to'
directive), we need to use the correct URL for the report, since that
URL's origin will be used to match the reporting endpoints. Otherwise,
we end up not sending reports or, worse, sending them to the wrong
endpoints.

This change addresses that and adds a WPT.

In the future, we should avoid sending the reports through Blink at
all.

Bug: 2324115
Change-Id: Iccab0cfe178900844d97aa7989922b014d02efcf